Add support for using lychee as a library
This splits up the code into a `lib` and a `bin`
to make the runtime usable from other crates.
