+++ Cargo.toml
@@ -4,6 +4,13 @@ description = "A glorious link checker"
 documentation = "https://github.com/lycheeverse/lychee/blob/master/README.md"
 edition = "2018"
 homepage = "https://github.com/lycheeverse/lychee"
+keywords = [
+    "link",
+    "checker",
+    "cli",
+    "link-checker",
+    "validator",
+]
 license = "Apache-2.0/MIT"
 name = "lychee"
 repository = "https://github.com/lycheeverse/lychee"
@@ -16,8 +23,6 @@ glob = "0.3"
 http = "0.2"
 hubcaps = "0.6"
 linkify = "0.4.0"
-log = "0.4"
-pretty_env_logger = "0.4"
 regex = "1.3.9"
 url = "2.1.1"
 check-if-email-exists = "0.8.13"
