+++ src/excludes.rs
@@ -1,0 +1,26 @@
+use regex::RegexSet;
+
+/// Exclude configuration for the link checker.
+/// You can ignore links based on
+#[derive(Clone, Debug)]
+pub struct Excludes {
+    pub regex: Option<RegexSet>,
+    /// Example: 192.168.0.1
+    pub private_ips: bool,
+    /// Example: 169.254.0.0
+    pub link_local_ips: bool,
+    /// For IPv4: 127.0. 0.1/8
+    /// For IPv6: ::1/128
+    pub loopback_ips: bool,
+}
+
+impl Default for Excludes {
+    fn default() -> Self {
+        Self {
+            regex: None,
+            private_ips: false,
+            link_local_ips: false,
+            loopback_ips: false,
+        }
+    }
+}
